Initial commit of v1.16.6
This is a minor update to v1.16.5 that will only be released in source code
form. It has some minor changes to v1.16.5 that have no effect on the functionality
of the plugin.

If any subsequent releases of the plugin will be made, they will be released both
in this repository in source code form and as a JAR on the website of the
corresponding Spigot resource, but no such releases will be made after September 30, 2020.